[feature] add and remove to favorite (#6)
* [feature] add and remove to favorite

* [feature] detekt fix

* [feature] detekt fix

+++ shared-test/src/main/java/com/santimattius/test/data/MovieMother.kt
@@ -29,14 +29,14 @@ object MovieMother {
     }
 
     fun createMovie(
-        id: Int = 508947,
+        id: Long = 508947,
         title: String = "Spider-Man: No Way Home",
         overview: String = "Spider-Man: No Way Home",
         poster: String = "/fOy2Jurz9k6RnJnMUMRDAgBwru2.jpg"
     ) = Movie(id, overview, title, poster)
 
     fun buildMovie(
-        id: Int = 508947,
+        id: Long = 508947,
         title: String = "Spider-Man: No Way Home",
         overview: String = "Spider-Man: No Way Home",
         poster: String = "/fOy2Jurz9k6RnJnMUMRDAgBwru2.jpg"
